Move `spawn_frame_base` into its own header/source pair.
