+++ CMakeLists.txt
@@ -8,6 +8,7 @@ find_package(Threads REQUIRED)
 set(Sources
 src/profiling.cpp
 src/spawn.cpp
+src/spawn_frame_base.cpp
 src/bulk_spawn.cpp
 src/this_thread.cpp
 src/thread_info.cpp
